Debug Endpoint and move application code to right assembly
